fix(InfiniteGrid): split infinitegrid for each layout.
