+++ src/index.framelayout.js
@@ -2,10 +2,9 @@
  * Copyright (c) NAVER Corp.
  * egjs-infinitegrid projects are licensed under the MIT license
  */
-import InfiniteGrid, * as modules from "./index";
+import InfiniteGrid from "./InfiniteGrid";
+import FrameLayout from "./layouts/FrameLayout";
 
-for (const name in modules) {
-	InfiniteGrid[name] = modules[name];
-}
+InfiniteGrid.FrameLayout = FrameLayout;
 
 export default InfiniteGrid;
