+++ src/index.justifiedlayout.js
@@ -1,0 +1,10 @@
+/**
+ * Copyright (c) NAVER Corp.
+ * egjs-infinitegrid projects are licensed under the MIT license
+ */
+import InfiniteGrid from "./InfiniteGrid";
+import JustifiedLayout from "./layouts/JustifiedLayout";
+
+InfiniteGrid.JustifiedLayout = JustifiedLayout;
+
+export default InfiniteGrid;
