+++ src/index.packinglayout.js
@@ -1,0 +1,10 @@
+/**
+ * Copyright (c) NAVER Corp.
+ * egjs-infinitegrid projects are licensed under the MIT license
+ */
+import InfiniteGrid from "./InfiniteGrid";
+import PackingLayout from "./layouts/PackingLayout";
+
+InfiniteGrid.PackingLayout = PackingLayout;
+
+export default InfiniteGrid;
