+++ src/index.squarelayout.js
@@ -1,0 +1,10 @@
+/**
+ * Copyright (c) NAVER Corp.
+ * egjs-infinitegrid projects are licensed under the MIT license
+ */
+import InfiniteGrid from "./InfiniteGrid";
+import SquareLayout from "./layouts/SquareLayout";
+
+InfiniteGrid.SquareLayout = SquareLayout;
+
+export default InfiniteGrid;
