--- src/version.js
@@ -1,11 +1,0 @@
-/**
- * Version info string
- * @ko 버전정보 문자열
- * @name VERSION
- * @static
- * @type {String}
- * @example
- * eg.InfiniteGrid.VERSION;  // ex) 3.3.3
- * @memberof eg.InfiniteGrid
- */
-module.exports = "#__VERSION__#";
